Add the first learning object.
This is an java object for learning the basic knowledge of the thread. It tells some use of the keyword 'synchronized' containing some details。What's more, it lets me know a new keyward called 'volatile' what can make variables(变量) visible(可见的) to muliple threads.However, 'volatile' is not atomicity, so it can't replace 'synchronized' in some scenarios that require synchronization.