Add the second learning object.
This Java object describe the use of 'wait' and 'notify'.They must be used with another keyword 'synchronized'。And 'wait' will release(释放) the rock, but 'notify' won't. So when we use 'wait' and 'notify', we must pay attention to feature(特征).
What's more,it tells me a new API:java.util.concurrent.CountDownLatch, which can realize(实现) the real-time communication(实时通信) in threads but no using 'wait' and 'notify'.
Then, we use the 'wait' and 'notify' to simulate(模拟，仿真) the blocking queue. And I know the 'ThreadLoacl'. It is a local variable(局部变量) of the thread, but it doesn't give the lock to ensure(确保) the threads are save. It use space to change time. The meaing is gving us a new idea to address thread-safety issues(解决线程安全问题).
Finally, this object tells me how to implement singleton pattern(单例模式).